# backend/src/Api/Controllers/ClinicalController.cs
using FamilyVeda.Application.Clinical;
using FamilyVeda.Application.Common;
using FamilyVeda.Application.Triage;
using FamilyVeda.Domain.Common;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;

namespace FamilyVeda.Api.Controllers;

[Route("api/v1")]
public sealed class ClinicalController(IClinicalService clinicalService) : ApiControllerBase
{
    [HttpPost("doctors/register")]
    [Authorize(Policy = "Doctor")]
    public async Task<ActionResult<DoctorDto>> RegisterDoctor(RegisterDoctorRequest request, CancellationToken cancellationToken)
    {
        var doctor = await clinicalService.RegisterDoctorAsync(request, cancellationToken);
        return Created("/api/v1/doctors/me", doctor);
    }

    [HttpGet("doctors/me")]
    [Authorize(Policy = "Doctor")]
    [Authorize(Policy = "Doctor")]
    public async Task<ActionResult<DoctorDto>> GetDoctor(CancellationToken cancellationToken) =>
        Ok(await clinicalService.GetMyDoctorAsync(cancellationToken));

    [HttpGet("doctors/me/cases")]
    [Authorize(Policy = "Doctor")]
    [Authorize(Policy = "Doctor")]
    public async Task<ActionResult<PagedResult<TriageCaseDto>>> GetMyCases(int page = 1, int pageSize = 20, CancellationToken cancellationToken = default) =>
        Ok(await clinicalService.GetMyCasesAsync(page, pageSize, cancellationToken));

    [HttpGet("doctors/case-pool")]
    [Authorize(Policy = "Doctor")]
    public async Task<ActionResult<PagedResult<AvailableCaseDto>>> GetCasePool(int page = 1, int pageSize = 20, CancellationToken cancellationToken = default) =>
        Ok(await clinicalService.GetAvailableCasesAsync(page, pageSize, cancellationToken));

    [HttpGet("admin/doctors")]
    [Authorize(Policy = "Admin")]
    [Authorize(Policy = "Admin")]
    public async Task<ActionResult<PagedResult<DoctorDto>>> GetPendingDoctors(int page = 1, int pageSize = 20, CancellationToken cancellationToken = default) =>
        Ok(await clinicalService.GetPendingDoctorsAsync(page, pageSize, cancellationToken));

    [HttpPost("admin/doctors/{doctorId:guid}/verification")]
    [Authorize(Policy = "Admin")]
    [Authorize(Policy = "Admin")]
    public async Task<ActionResult<DoctorDto>> ChangeVerification(Guid doctorId, VerifyDoctorRequest request, CancellationToken cancellationToken) =>
        Ok(await clinicalService.ChangeVerificationAsync(doctorId, request, cancellationToken));

    [HttpPost("admin/doctors/{doctorId:guid}/verify")]
    [Authorize(Policy = "Admin")]
    public async Task<ActionResult<DoctorDto>> VerifyDoctor(Guid doctorId, VerificationReasonRequest request, CancellationToken cancellationToken) =>
        Ok(await clinicalService.ChangeVerificationAsync(doctorId, new VerifyDoctorRequest(VerificationStatus.Verified, request.Reason), cancellationToken));

    [HttpPost("admin/doctors/{doctorId:guid}/request-info")]
    [Authorize(Policy = "Admin")]
    public async Task<ActionResult<DoctorDto>> RequestDoctorInfo(Guid doctorId, VerificationReasonRequest request, CancellationToken cancellationToken) =>
        Ok(await clinicalService.ChangeVerificationAsync(doctorId, new VerifyDoctorRequest(VerificationStatus.MoreInformationRequired, request.Reason), cancellationToken));

    [HttpPost("admin/doctors/{doctorId:guid}/reject")]
    [Authorize(Policy = "Admin")]
    public async Task<ActionResult<DoctorDto>> RejectDoctor(Guid doctorId, VerificationReasonRequest request, CancellationToken cancellationToken) =>
        Ok(await clinicalService.ChangeVerificationAsync(doctorId, new VerifyDoctorRequest(VerificationStatus.Rejected, request.Reason), cancellationToken));

    [HttpPost("admin/doctors/{doctorId:guid}/suspend")]
    [Authorize(Policy = "Admin")]
    public async Task<ActionResult<DoctorDto>> SuspendDoctor(Guid doctorId, VerificationReasonRequest request, CancellationToken cancellationToken) =>
        Ok(await clinicalService.ChangeVerificationAsync(doctorId, new VerifyDoctorRequest(VerificationStatus.Suspended, request.Reason), cancellationToken));

    [HttpPost("triage-cases/{caseId:guid}/claim")]
    [Authorize(Policy = "Doctor")]
    [Authorize(Policy = "Doctor")]
    public async Task<ActionResult<ApprovalDto>> Claim(Guid caseId, CancellationToken cancellationToken) =>
        Ok(await clinicalService.ClaimCaseAsync(caseId, cancellationToken));

    [HttpPost("triage-cases/{caseId:guid}/decision")]
    [Authorize(Policy = "Doctor")]
    public async Task<ActionResult<ApprovalDto>> Decide(Guid caseId, ApprovalRequest request, CancellationToken cancellationToken) =>
        Ok(await clinicalService.DecideCaseAsync(caseId, request, cancellationToken));

    [HttpPost("triage-cases/{caseId:guid}/approve")]
    [Authorize(Policy = "Doctor")]
    public async Task<ActionResult<ApprovalDto>> Approve(Guid caseId, ApprovalContentRequest request, CancellationToken cancellationToken) =>
        Ok(await clinicalService.DecideCaseAsync(caseId, new ApprovalRequest(ApprovalAction.Approve, request.DoctorNotes, request.FinalAdvisory), cancellationToken));

    [HttpPost("triage-cases/{caseId:guid}/revise")]
    [Authorize(Policy = "Doctor")]
    public async Task<ActionResult<ApprovalDto>> Revise(Guid caseId, ApprovalContentRequest request, CancellationToken cancellationToken) =>
        Ok(await clinicalService.DecideCaseAsync(caseId, new ApprovalRequest(ApprovalAction.ReviseAndApprove, request.DoctorNotes, request.FinalAdvisory), cancellationToken));

    [HttpPost("triage-cases/{caseId:guid}/request-info")]
    [Authorize(Policy = "Doctor")]
    public async Task<ActionResult<ApprovalDto>> RequestCaseInfo(Guid caseId, ApprovalContentRequest request, CancellationToken cancellationToken) =>
        Ok(await clinicalService.DecideCaseAsync(caseId, new ApprovalRequest(ApprovalAction.RequestInformation, request.DoctorNotes, null), cancellationToken));

    [HttpPost("triage-cases/{caseId:guid}/reject")]
    [Authorize(Policy = "Doctor")]
    public async Task<ActionResult<ApprovalDto>> Reject(Guid caseId, ApprovalContentRequest request, CancellationToken cancellationToken) =>
        Ok(await clinicalService.DecideCaseAsync(caseId, new ApprovalRequest(ApprovalAction.Reject, request.DoctorNotes, null), cancellationToken));

    [HttpPost("triage-cases/{caseId:guid}/escalate")]
    [Authorize(Policy = "Doctor")]
    public async Task<ActionResult<ApprovalDto>> Escalate(Guid caseId, ApprovalContentRequest request, CancellationToken cancellationToken) =>
        Ok(await clinicalService.DecideCaseAsync(caseId, new ApprovalRequest(ApprovalAction.Escalate, request.DoctorNotes, null), cancellationToken));

    [HttpGet("audit")]
    [Authorize(Roles = "Admin,FamilyUser")]
    public async Task<ActionResult<PagedResult<AuditDto>>> GetAudit(Guid? subjectMemberId = null, int page = 1, int pageSize = 20, CancellationToken cancellationToken = default) =>
        Ok(await clinicalService.GetAuditAsync(subjectMemberId, page, pageSize, cancellationToken));
}