# docs/api/openapi.yaml
...
openapi: 3.0.1
info:
  title: Family Veda API
  version: v1
paths:
  "/api/v1/auth/register":
    post:
      tags:
      - Auth
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/RegisterRequest"
          text/json:
            schema:
              "$ref": "#/components/schemas/RegisterRequest"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/RegisterRequest"
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AuthResponse"
  "/api/v1/auth/login":
    post:
      tags:
      - Auth
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/LoginRequest"
          text/json:
            schema:
              "$ref": "#/components/schemas/LoginRequest"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/LoginRequest"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AuthResponse"
  "/api/v1/auth/refresh":
    post:
      tags:
      - Auth
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/RefreshRequest"
          text/json:
            schema:
              "$ref": "#/components/schemas/RefreshRequest"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/RefreshRequest"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AuthResponse"
...
  "/api/v1/doctors/register":
    post:
      tags:
      - Clinical
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/RegisterDoctorRequest"
          text/json:
            schema:
              "$ref": "#/components/schemas/RegisterDoctorRequest"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/RegisterDoctorRequest"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/DoctorDto"
...
  "/api/v1/doctors/me/cases":
    get:
      tags:
      - Clinical
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          format: int32
          default: 1
      - name: pageSize
        in: query
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TriageCaseDtoPagedResult"
  "/api/v1/doctors/case-pool":
    get:
      tags:
      - Clinical
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          format: int32
          default: 1
      - name: pageSize
        in: query
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AvailableCaseDtoPagedResult"
  "/api/v1/admin/doctors":
    get:
      tags:
      - Clinical
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          format: int32
          default: 1
      - name: pageSize
        in: query
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/DoctorDtoPagedResult"
  "/api/v1/admin/doctors/{doctorId}/verification":
    post:
      tags:
      - Clinical
      parameters:
      - name: doctorId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/VerifyDoctorRequest"
          text/json:
            schema:
              "$ref": "#/components/schemas/VerifyDoctorRequest"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/VerifyDoctorRequest"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/DoctorDto"
  "/api/v1/admin/doctors/{doctorId}/verify":
    post:
      tags:
      - Clinical
      parameters:
      - name: doctorId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/VerificationReasonRequest"
          text/json:
            schema:
              "$ref": "#/components/schemas/VerificationReasonRequest"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/VerificationReasonRequest"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/DoctorDto"
  "/api/v1/admin/doctors/{doctorId}/request-info":
    post:
      tags:
      - Clinical
      parameters:
      - name: doctorId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/VerificationReasonRequest"
          text/json:
            schema:
              "$ref": "#/components/schemas/VerificationReasonRequest"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/VerificationReasonRequest"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/DoctorDto"
  "/api/v1/admin/doctors/{doctorId}/reject":
    post:
      tags:
      - Clinical
      parameters:
      - name: doctorId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/VerificationReasonRequest"
          text/json:
            schema:
              "$ref": "#/components/schemas/VerificationReasonRequest"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/VerificationReasonRequest"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/DoctorDto"
  "/api/v1/admin/doctors/{doctorId}/suspend":
    post:
      tags:
      - Clinical
      parameters:
      - name: doctorId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/VerificationReasonRequest"
          text/json:
            schema:
              "$ref": "#/components/schemas/VerificationReasonRequest"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/VerificationReasonRequest"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/DoctorDto"
  "/api/v1/triage-cases/{caseId}/claim":
    post:
      tags:
      - Clinical
      parameters:
      - name: caseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ApprovalDto"
  "/api/v1/triage-cases/{caseId}/decision":
    post:
      tags:
      - Clinical
      parameters:
      - name: caseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ApprovalRequest"
          text/json:
            schema:
              "$ref": "#/components/schemas/ApprovalRequest"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/ApprovalRequest"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ApprovalDto"
  "/api/v1/triage-cases/{caseId}/approve":
    post:
      tags:
      - Clinical
      parameters:
      - name: caseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ApprovalContentRequest"
          text/json:
            schema:
              "$ref": "#/components/schemas/ApprovalContentRequest"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/ApprovalContentRequest"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ApprovalDto"
  "/api/v1/triage-cases/{caseId}/revise":
    post:
      tags:
      - Clinical
      parameters:
      - name: caseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ApprovalContentRequest"
          text/json:
            schema:
              "$ref": "#/components/schemas/ApprovalContentRequest"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/ApprovalContentRequest"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ApprovalDto"
  "/api/v1/triage-cases/{caseId}/request-info":
    post:
      tags:
      - Clinical
      parameters:
      - name: caseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ApprovalContentRequest"
          text/json:
            schema:
              "$ref": "#/components/schemas/ApprovalContentRequest"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/ApprovalContentRequest"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ApprovalDto"
  "/api/v1/triage-cases/{caseId}/reject":
    post:
      tags:
      - Clinical
      parameters:
      - name: caseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ApprovalContentRequest"
          text/json:
            schema:
              "$ref": "#/components/schemas/ApprovalContentRequest"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/ApprovalContentRequest"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ApprovalDto"
  "/api/v1/triage-cases/{caseId}/escalate":
    post:
      tags:
      - Clinical
      parameters:
      - name: caseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ApprovalContentRequest"
          text/json:
            schema:
              "$ref": "#/components/schemas/ApprovalContentRequest"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/ApprovalContentRequest"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ApprovalDto"
  "/api/v1/audit":
    get:
      tags:
      - Clinical
      parameters:
      - name: subjectMemberId
        in: query
        schema:
          type: string
          format: uuid
      - name: page
        in: query
        schema:
          type: integer
          format: int32
          default: 1
      - name: pageSize
        in: query
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AuditDtoPagedResult"
...
components:
  schemas:
...
    ApprovalAction:
      enum:
      - Approve
      - ReviseAndApprove
      - RequestInformation
      - Reject
      - Escalate
      type: string
    ApprovalContentRequest:
      type: object
      properties:
        doctorNotes:
          type: string
          nullable: true
        finalAdvisory:
          type: string
          nullable: true
      additionalProperties: false
    ApprovalDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        triageCaseId:
          type: string
          format: uuid
        doctorId:
          type: string
          format: uuid
        action:
          "$ref": "#/components/schemas/ApprovalAction"
        decidedAt:
          type: string
          format: date-time
      additionalProperties: false
    ApprovalRequest:
      type: object
      properties:
        action:
          "$ref": "#/components/schemas/ApprovalAction"
        doctorNotes:
          type: string
          nullable: true
        finalAdvisory:
          type: string
          nullable: true
      additionalProperties: false
...
    AuditDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        eventType:
          type: string
          nullable: true
        resourceType:
          type: string
          nullable: true
        resourceId:
          type: string
          format: uuid
          nullable: true
        outcome:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      additionalProperties: false
    AuditDtoPagedResult:
      type: object
      properties:
        items:
          type: array
          items:
            "$ref": "#/components/schemas/AuditDto"
          nullable: true
        page:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalCount:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
          readOnly: true
      additionalProperties: false
    AuthResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        displayName:
          type: string
          nullable: true
        userType:
          "$ref": "#/components/schemas/UserType"
        familyRole:
          "$ref": "#/components/schemas/FamilyRole"
        doctorVerificationStatus:
          "$ref": "#/components/schemas/VerificationStatus"
        accessToken:
          type: string
          nullable: true
        refreshToken:
          type: string
          nullable: true
        accessTokenExpiresAt:
          type: string
          format: date-time
      additionalProperties: false
    AvailableCaseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        priority:
          "$ref": "#/components/schemas/TriagePriority"
        createdAt:
          type: string
          format: date-time
      additionalProperties: false
    AvailableCaseDtoPagedResult:
      type: object
      properties:
        items:
          type: array
          items:
            "$ref": "#/components/schemas/AvailableCaseDto"
          nullable: true
        page:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalCount:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
          readOnly: true
      additionalProperties: false
...
    DoctorDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        registrationNumberLastFour:
          type: string
          nullable: true
        verificationStatus:
          "$ref": "#/components/schemas/VerificationStatus"
        specialty:
          type: string
          nullable: true
      additionalProperties: false
    DoctorDtoPagedResult:
      type: object
      properties:
        items:
          type: array
          items:
            "$ref": "#/components/schemas/DoctorDto"
          nullable: true
        page:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalCount:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
          readOnly: true
      additionalProperties: false
...
    FamilyRole:
      enum:
      - Head
      - AdultMember
      - MinorMember
      type: string
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
      additionalProperties: false
...
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
          nullable: true
      additionalProperties: false
    RegisterDoctorRequest:
      type: object
      properties:
        registrationNumber:
          type: string
          nullable: true
        specialty:
          type: string
          nullable: true
      additionalProperties: false
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
        displayName:
          type: string
          nullable: true
        userType:
          "$ref": "#/components/schemas/UserType"
      additionalProperties: false
...
    TriageCaseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        episodeId:
          type: string
          format: uuid
        memberId:
          type: string
          format: uuid
        status:
          "$ref": "#/components/schemas/TriageStatus"
        priority:
          "$ref": "#/components/schemas/TriagePriority"
        createdAt:
          type: string
          format: date-time
      additionalProperties: false
    TriageCaseDtoPagedResult:
      type: object
      properties:
        items:
          type: array
          items:
            "$ref": "#/components/schemas/TriageCaseDto"
          nullable: true
        page:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalCount:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
          readOnly: true
      additionalProperties: false
    TriagePriority:
      enum:
      - Routine
      - Priority
      - Emergency
      type: string
    TriageStatus:
      enum:
      - Submitted
      - Planning
      - ContextReady
      - Analysed
      - RiskAssessed
      - Validated
      - LowConfidence
      - PendingDoctorReview
      - Claimed
      - Approved
      - ApprovedRevised
      - Rejected
      - Escalated
      - FailedSafe
      type: string
...
    UserType:
      enum:
      - FamilyUser
      - Doctor
      - Admin
      type: string
    VerificationReasonRequest:
      type: object
      properties:
        reason:
          type: string
          nullable: true
      additionalProperties: false
    VerificationStatus:
      enum:
      - Pending
      - Verified
      - MoreInformationRequired
      - Rejected
      - Suspended
      type: string
    VerifyDoctorRequest:
      type: object
      properties:
        status:
          "$ref": "#/components/schemas/VerificationStatus"
        reason:
          type: string
          nullable: true
      additionalProperties: false
...
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
- Bearer: []